chore: Add customElements to Jest Environment
